+++ modules/entities/ts/item/interfaces/config.ts
@@ -1,8 +1,12 @@
+import { TAdapters } from '../../adapter';
 import { IProvider } from '../../interfaces/provider';
 
-export interface IItemConfig {
+export interface IItemConfig<T = any> {
 	storeName?: string;
 	db?: string;
 	id?: string | number;
-	provider?: new () => IProvider;
+	localdb?: boolean;
+	properties?: { [key: string]: T };
+	provider?: new (item?: any) => IProvider;
+	adapter?: TAdapters;
 }