+++ modules/entities/ts/item/save.ts
@@ -1,16 +1,21 @@
+import type { ResponseAdapter } from '../adapter';
+import { IResponseAdapter } from '../adapter/interface';
 import type { Item } from './index';
+import type { LocalProvider } from './local-provider';
 
 export class ItemSaveManager {
 	#parent: Item<any>;
 	#getProperty;
 	#bridge;
 	#provider;
-	#localProvider;
+	#localProvider: LocalProvider;
 
+	#adapter: IResponseAdapter;
 	constructor(parent: Item<any>, bridge) {
 		this.#parent = parent;
 		this.#getProperty = bridge.get;
 		this.#bridge = bridge;
+		this.#adapter = this.#parent.responseAdapter;
 		this.init();
 	}
 
@@ -18,10 +23,9 @@ export class ItemSaveManager {
 		this.#parent.localUpdate = this.localUpdate;
 		this.#localProvider = this.#getProperty('localProvider');
 		this.#provider = this.#getProperty('provider');
-		this.#parent.sync = this.sync;
 	}
 
-	save = async (data = undefined) => {
+	save = async (data?) => {
 		try {
 			await this.#getProperty('checkReady')();
 
@@ -29,22 +33,26 @@ export class ItemSaveManager {
 				this.#parent.set(data);
 			}
 
-			if (!this.#parent.isUnpublished) {
-				return;
-			}
+			if (!this.#parent.isUnpublished) return;
 
-			const properties = this.#parent.getProperties();
+			const properties = { ...data, ...this.#parent.getProperties() };
+			properties.isNew = this.#localProvider.registry.isNew;
+
+			if (this.#parent.isOnline) {
+				const response = await this.#publish(properties);
+				this.#adapter.fromRemote(response);
+				this.#localProvider.registry.isNew = false;
+			}
 
 			if (this.#localProvider) {
 				await this.#localProvider.save(properties);
 			}
-
-			await this.#publish(properties);
 			this.#parent.triggerEvent();
 
-			return { status: true };
+			return this.#adapter.toClient();
 		} catch (e) {
 			console.error('error saving', e);
+			return e;
 		}
 	};
 
@@ -53,17 +61,16 @@ export class ItemSaveManager {
 			if (!this.#provider || !this.#bridge.get('isOnline')) return;
 
 			const response = await this.#provider.publish(properties);
-
-			if (!response?.status) throw response.error;
+			const data = this.#adapter.fromRemote(response);
 
 			if (this.#localProvider) {
-				this.#localProvider.save(response.data, true);
+				this.#localProvider.save(data, true);
 				this.#localProvider.triggerEvent();
 			}
-			return { status: true, data: response };
+			return this.#adapter.toClient({ data });
 		} catch (error) {
 			console.error('ERROR PUBLISHING', error);
-			return { status: false, error };
+			return this.#adapter.toClient({ error });
 		}
 	};
 
@@ -99,7 +106,7 @@ export class ItemSaveManager {
 
 			this.#parent.triggerEvent();
 
-			return { status: true };
+			return this.#adapter.toClient();
 		} catch (e) {
 			console.error('error updating locally', e);
 		}